Add endpoints for retrieving expression vectors for a given gene
Rename some of the endpoints that retrieve expression levels to be
consistent with the pluralization of their parameters.

Relocate methods that retrieves vectors from cache in a separate
service.

Optimize some of our queries to handle large number of datasets.

Add trace logging to QueryUtils for listByBatch() and listByIdentifiableBatch().
