+++ gemma-core/src/main/java/ubic/gemma/persistence/service/expression/bioAssayData/AbstractDesignElementDataVectorDao.java
@@ -63,6 +63,10 @@ public abstract class AbstractDesignElementDataVectorDao<T extends DesignElement
 
     @Override
     public void thaw( Collection<T> designElementDataVectors ) {
+        if ( designElementDataVectors.isEmpty() ) {
+            return;
+        }
+
         StopWatch timer = StopWatch.createStarted();
         StopWatch vTimer = StopWatch.create(),
                 eeTimer = StopWatch.create(),
