+++ gemma-core/src/test/java/ubic/gemma/core/analysis/preprocess/VectorMergingServiceTest.java
@@ -17,7 +17,6 @@ package ubic.gemma.core.analysis.preprocess;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -247,7 +246,7 @@ public class VectorMergingServiceTest extends AbstractGeoServiceTest {
         assertEquals( 46, ee.getNumberOfSamples().intValue() );
 
         Collection<DoubleVectorValueObject> processedDataArrays = processedExpressionDataVectorService
-                .getProcessedDataArrays( ee, 50 );
+                .getRandomProcessedDataArrays( ee, 50 );
 
         assertEquals( 28, processedDataArrays.size() );
