+++ gemma-rest/src/main/java/ubic/gemma/rest/DatasetsWebService.java
@@ -86,10 +86,10 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import javax.ws.rs.*;
 import javax.ws.rs.core.*;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.net.URI;
+import java.nio.file.Files;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -214,12 +214,7 @@ public class DatasetsWebService {
             ids.sort( Comparator.comparingDouble( i -> -scoreById.get( i ) ) );
 
             // slice the ranked IDs
-            List<Long> idsSlice;
-            if ( offset < ids.size() ) {
-                idsSlice = ids.subList( offset, Math.min( offset + limit, ids.size() ) );
-            } else {
-                idsSlice = Collections.emptyList();
-            }
+            List<Long> idsSlice = sliceIds( ids, offset, limit );
 
             // now highlight the results in the slice
             List<SearchResult<ExpressionExperiment>> results = datasetArgService.getResultsForSearchQuery( query, new Highlighter( new HashSet<>( idsSlice ) ) );
@@ -702,13 +697,20 @@ public class DatasetsWebService {
                     content = @Content(schema = @Schema(implementation = ResponseErrorObject.class))) })
     public ResponseDataObject<List<DifferentialExpressionAnalysisValueObject>> getDatasetDifferentialExpressionAnalyses( // Params:
             @PathParam("dataset") DatasetArg<?> datasetArg, // Required
-            @QueryParam("offset") @DefaultValue("0") OffsetArg offset, // Optional, default 0
-            @QueryParam("limit") @DefaultValue("20") LimitArg limit // Optional, default 20
+            @QueryParam("offset") @DefaultValue("0") OffsetArg offsetArg, // Optional, default 0
+            @QueryParam("limit") @DefaultValue("20") LimitArg limitArg // Optional, default 20
     ) {
-        return respond(
-                this.getDiffExVos( datasetArgService.getEntity( datasetArg ).getId(),
-                        offset.getValue(), limit.getValue() )
-        );
+        List<DifferentialExpressionAnalysisValueObject> result;
+        Long eeId = datasetArgService.getEntity( datasetArg ).getId();
+        int offset = offsetArg.getValue();
+        int limit = limitArg.getValue();
+        Map<ExpressionExperimentDetailsValueObject, List<DifferentialExpressionAnalysisValueObject>> map = differentialExpressionAnalysisService.getAnalysesByExperiment( Collections.singleton( eeId ), offset, limit );
+        if ( map == null || map.isEmpty() ) {
+            result = Collections.emptyList();
+        } else {
+            result = map.get( map.keySet().iterator().next() );
+        }
+        return respond( result );
     }
 
     /**
@@ -750,7 +752,7 @@ public class DatasetsWebService {
     @Path("/analyses/differential/results/genes/{gene}")
     @Produces({ MediaType.APPLICATION_JSON, MediaTypeUtils.TEXT_TAB_SEPARATED_VALUES_UTF8 })
     @Operation(
-            summary = "Retrieve the differential expression results for a given gene",
+            summary = "Retrieve the differential expression results for a given gene among datasets matching the provided query and filter",
             description = GET_DATASETS_DIFFERENTIAL_ANALYSIS_EXPRESSION_RESULTS_DESCRIPTION,
             responses = {
                     @ApiResponse(content = {
@@ -787,7 +789,7 @@ public class DatasetsWebService {
     @Path("/analyses/differential/results/taxa/{taxon}/genes/{gene}")
     @Produces({ MediaType.APPLICATION_JSON, MediaTypeUtils.TEXT_TAB_SEPARATED_VALUES_UTF8 })
     @Operation(
-            summary = "Retrieve the differential expression results for a given gene and taxa",
+            summary = "Retrieve the differential expression results for a given gene and taxa among datasets matching the provided query and filter",
             description = GET_DATASETS_DIFFERENTIAL_ANALYSIS_EXPRESSION_RESULTS_DESCRIPTION,
             responses = {
                     @ApiResponse(content = {
@@ -826,19 +828,18 @@ public class DatasetsWebService {
         if ( threshold < 0 || threshold > 1 ) {
             throw new BadRequestException( "The threshold must be in the [0, 1] interval." );
         }
-        Set<Long> ids = new HashSet<>( expressionExperimentService.loadIdsWithCache( filters, expressionExperimentService.getSort( "id", Sort.Direction.ASC ) ) );
+        List<Long> ids = new ArrayList<>( expressionExperimentService.loadIdsWithCache( filters, expressionExperimentService.getSort( "id", Sort.Direction.ASC ) ) );
         if ( query != null ) {
             ids.retainAll( datasetArgService.getIdsForSearchQuery( query ) );
         }
         // slice IDs
         long totalElements = ids.size();
-        ids = ids.stream().sorted().skip( offset ).limit( limit ).collect( Collectors.toSet() );
+        List<Long> slicedIds = sliceIds( ids, offset, limit );
         Map<DifferentialExpressionAnalysisResult, Long> sourceExperimentIdMap = new HashMap<>();
         Map<DifferentialExpressionAnalysisResult, Long> experimentAnalyzedIdMap = new HashMap<>();
         Map<DifferentialExpressionAnalysisResult, Baseline> baselineMap = new HashMap<>();
         List<DifferentialExpressionAnalysisResultByGeneValueObject> payload = differentialExpressionResultService.findByGeneAndExperimentAnalyzed( gene, ids, sourceExperimentIdMap, experimentAnalyzedIdMap, baselineMap, threshold, false ).stream()
                 .map( r -> new DifferentialExpressionAnalysisResultByGeneValueObject( r, sourceExperimentIdMap.get( r ), experimentAnalyzedIdMap.get( r ), baselineMap.get( r ) ) )
-                // because of pagination, this is the most "adequate" order
                 .sorted( Comparator.comparing( DifferentialExpressionAnalysisResultByGeneValueObject::getSourceExperimentId )
                         .thenComparing( DifferentialExpressionAnalysisResultByGeneValueObject::getExperimentAnalyzedId )
                         .thenComparing( DifferentialExpressionAnalysisResultByGeneValueObject::getResultSetId ) )
@@ -989,7 +990,12 @@ public class DatasetsWebService {
     ) {
         ExpressionExperiment ee = datasetArgService.getEntity( datasetArg );
         try {
-            return this.outputDataFile( ee, filterData );
+            ee = expressionExperimentService.thawLite( ee );
+            File file = expressionDataFileService.writeOrLocateProcessedDataFile( ee, false, filterData ).orElse( null );
+            if ( file == null || !file.exists() ) {
+                throw new NotFoundException( String.format( DatasetsWebService.ERROR_DATA_FILE_NOT_AVAILABLE, ee.getShortName() ) );
+            }
+            return this.outputFile( file );
         } catch ( NoRowsLeftAfterFilteringException e ) {
             return Response.noContent().build();
         } catch ( FilteringException e ) {
@@ -1077,7 +1083,12 @@ public class DatasetsWebService {
     ) {
         ExpressionExperiment ee = datasetArgService.getEntity( datasetArg );
         try {
-            return this.outputDesignFile( ee );
+            ee = expressionExperimentService.thawLite( ee );
+            File file = expressionDataFileService.writeOrLocateDesignFile( ee, false );
+            if ( file == null || !file.exists() ) {
+                throw new NotFoundException( String.format( DatasetsWebService.ERROR_DESIGN_FILE_NOT_AVAILABLE, ee.getShortName() ) );
+            }
+            return this.outputFile( file );
         } catch ( IOException e ) {
             throw new InternalServerErrorException( e.getMessage(), e );
         }
@@ -1218,6 +1229,65 @@ public class DatasetsWebService {
     }
 
     /**
+     * Retrieve the expression levels of a given gene across all datasets.
+     */
+    @GET
+    @Path("/expressions/genes/{gene}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Operation(summary = "Retrieve the expression levels of a gene among datasets matching the provided query and filter")
+    public PaginatedResponseDataObject<ExperimentExpressionLevelsValueObject> getDatasetsExpressionLevelsForGene(
+            @PathParam("gene") GeneArg<?> geneArg,
+            @QueryParam("query") QueryArg queryArg,
+            @QueryParam("filter") @DefaultValue("") FilterArg<ExpressionExperiment> filterArg,
+            @QueryParam("offset") @DefaultValue("0") OffsetArg offsetArg,
+            @QueryParam("limit") @DefaultValue("20") LimitArg limitArg,
+            @QueryParam("keepNonSpecific") @DefaultValue("false") Boolean keepNonSpecific, // Optional, default false
+            @QueryParam("consolidate") ExpLevelConsolidationArg consolidate // Optional, default everything is returned
+
+    ) {
+        return getDatasetsExpressionLevelsForGeneInTaxonInternal( geneArgService.getEntity( geneArg ), queryArg, filterArg, offsetArg, limitArg, keepNonSpecific, consolidate );
+    }
+
+    /**
+     * Retrieve the expression levels of a given gene and taxon across all datasets.
+     */
+    @GET
+    @Path("/expressions/taxa/{taxon}/genes/{gene}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Operation(summary = "Retrieve the expression levels of a gene and taxa among datasets matching the provided query and filter")
+    public PaginatedResponseDataObject<ExperimentExpressionLevelsValueObject> getDatasetsExpressionLevelsForGeneInTaxon(
+            @PathParam("taxon") TaxonArg<?> taxonArg,
+            @PathParam("gene") GeneArg<?> geneArg,
+            @QueryParam("query") QueryArg queryArg,
+            @QueryParam("filter") @DefaultValue("") FilterArg<ExpressionExperiment> filterArg,
+            @QueryParam("offset") @DefaultValue("0") OffsetArg offsetArg,
+            @QueryParam("limit") @DefaultValue("20") LimitArg limitArg,
+            @QueryParam("keepNonSpecific") @DefaultValue("false") Boolean keepNonSpecific, // Optional, default false
+            @QueryParam("consolidate") ExpLevelConsolidationArg consolidate // Optional, default everything is returned
+
+    ) {
+        return getDatasetsExpressionLevelsForGeneInTaxonInternal( geneArgService.getEntityWithTaxon( geneArg, taxonArgService.getEntity( taxonArg ) ), queryArg, filterArg, offsetArg, limitArg, keepNonSpecific, consolidate );
+    }
+
+    private QueriedAndFilteredAndInferredAndPaginatedResponseDataObject<ExperimentExpressionLevelsValueObject> getDatasetsExpressionLevelsForGeneInTaxonInternal( Gene gene, @Nullable QueryArg queryArg, FilterArg<ExpressionExperiment> filterArg, OffsetArg offsetArg, LimitArg limitArg, boolean keepNonSpecific, @Nullable ExpLevelConsolidationArg consolidate ) {
+        Collection<OntologyTerm> inferredTerms = new HashSet<>();
+        Filters filter = datasetArgService.getFilters( filterArg, null, inferredTerms );
+        Sort sort = datasetArgService.getSort( SortArg.valueOf( "+id" ) );
+        List<Long> datasetIds = expressionExperimentService.loadIdsWithCache( filter, sort );
+        if ( queryArg != null ) {
+            datasetIds.retainAll( datasetArgService.getIdsForSearchQuery( queryArg ) );
+        }
+        int offset = offsetArg.getValue();
+        int limit = limitArg.getValue();
+        Slice<ExperimentExpressionLevelsValueObject> slice = new Slice<>( processedExpressionDataVectorService
+                .getExpressionLevelsByIds( sliceIds( datasetIds, offset, limit ),
+                        Collections.singleton( gene ),
+                        keepNonSpecific,
+                        consolidate == null ? null : consolidate.getValue() ), sort, offset, limit, ( long ) datasetIds.size() );
+        return paginate( slice, queryArg != null ? queryArg.getValue() : null, filter, new String[] { "datasetId" }, inferredTerms );
+    }
+
+    /**
      * Retrieves the expression levels of given genes on given datasets.
      *
      * @param datasets        a list of dataset identifiers separated by commas (','). The identifiers can either be the
@@ -1251,12 +1321,12 @@ public class DatasetsWebService {
      *                        </ul>
      */
     @GET
-    @Path("/{datasets}/expressions/taxa/{taxa}/genes/{genes}")
+    @Path("/{datasets}/expressions/taxa/{taxon}/genes/{genes}")
     @Produces(MediaType.APPLICATION_JSON)
     @Operation(summary = "Retrieve the expression data matrix of a set of datasets and genes")
-    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetExpressionForGenesInTaxa( // Params:
+    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetsExpressionLevelsForGenesInTaxon( // Params:
             @PathParam("datasets") DatasetArrayArg datasets, // Required
-            @PathParam("taxa") TaxonArg<?> taxonArg, // Required
+            @PathParam("taxon") TaxonArg<?> taxonArg, // Required
             @PathParam("genes") GeneArrayArg genes, // Required
             @QueryParam("keepNonSpecific") @DefaultValue("false") Boolean keepNonSpecific, // Optional, default false
             @QueryParam("consolidate") ExpLevelConsolidationArg consolidate // Optional, default everything is returned
@@ -1273,7 +1343,7 @@ public class DatasetsWebService {
     @Path("/{datasets}/expressions/genes/{genes}")
     @Produces(MediaType.APPLICATION_JSON)
     @Operation(summary = "Retrieve the expression data matrix of a set of datasets and genes")
-    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetExpressionForGenes( // Params:
+    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetsExpressionLevelsForGenes( // Params:
             @PathParam("datasets") DatasetArrayArg datasets, // Required
             @PathParam("genes") GeneArrayArg genes, // Required
             @QueryParam("keepNonSpecific") @DefaultValue("false") Boolean keepNonSpecific, // Optional, default false
@@ -1313,7 +1383,7 @@ public class DatasetsWebService {
     @Path("/{datasets}/expressions/pca")
     @Produces(MediaType.APPLICATION_JSON)
     @Operation(summary = "Retrieve the principal components (PCA) of a set of datasets")
-    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetExpressionPca( // Params:
+    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetsExpressionPca( // Params:
             @PathParam("datasets") DatasetArrayArg datasets, // Required
             @QueryParam("component") @DefaultValue("1") Integer component, // Required, default 1
             @QueryParam("limit") @DefaultValue("100") LimitArg limit, // Optional, default 100
@@ -1355,7 +1425,7 @@ public class DatasetsWebService {
     @Path("/{datasets}/expressions/differential")
     @Produces(MediaType.APPLICATION_JSON)
     @Operation(summary = "Retrieve the expression levels of a set of datasets subject to a threshold on their differential expressions")
-    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetDifferentialExpression( // Params:
+    public ResponseDataObject<List<ExperimentExpressionLevelsValueObject>> getDatasetsDifferentialExpression( // Params:
             @PathParam("datasets") DatasetArrayArg datasets, // Required
             @QueryParam("diffExSet") Long diffExSet, // Required
             @QueryParam("threshold") @DefaultValue("1.0") Double threshold, // Optional, default 1.0
@@ -1413,40 +1483,16 @@ public class DatasetsWebService {
                 .build();
     }
 
-    private Response outputDataFile( ExpressionExperiment ee, boolean filter ) throws FilteringException, IOException {
-        ee = expressionExperimentService.thawLite( ee );
-        File file = expressionDataFileService.writeOrLocateProcessedDataFile( ee, false, filter ).orElse( null );
-        return this.outputFile( file, DatasetsWebService.ERROR_DATA_FILE_NOT_AVAILABLE, ee.getShortName() );
-    }
-
-    private Response outputDesignFile( ExpressionExperiment ee ) throws IOException {
-        ee = expressionExperimentService.thawLite( ee );
-        File file = expressionDataFileService.writeOrLocateDesignFile( ee, false );
-        return this.outputFile( file, DatasetsWebService.ERROR_DESIGN_FILE_NOT_AVAILABLE, ee.getShortName() );
-    }
-
-    private Response outputFile( @Nullable File file, String error, String shortName ) throws IOException {
-        if ( file == null || !file.exists() ) {
-            throw new NotFoundException( String.format( error, shortName ) );
-        }
+    private Response outputFile( File file ) throws IOException {
         // we remove the .gz extension because we use HTTP Content-Encoding
-        return Response.ok( new GZIPInputStream( new FileInputStream( file ) ) )
+        return Response.ok( new GZIPInputStream( Files.newInputStream( file.toPath() ) ) )
                 .header( "Content-Encoding", "gzip" )
                 .header( "Content-Disposition", "attachment; filename=" + FilenameUtils.removeExtension( file.getName() ) )
                 .build();
     }
 
-    private List<DifferentialExpressionAnalysisValueObject> getDiffExVos( Long eeId, int offset, int limit ) {
-        Map<ExpressionExperimentDetailsValueObject, List<DifferentialExpressionAnalysisValueObject>> map = differentialExpressionAnalysisService
-                .getAnalysesByExperiment( Collections.singleton( eeId ), offset, limit );
-        if ( map == null || map.isEmpty() ) {
-            return Collections.emptyList();
-        }
-        return map.get( map.keySet().iterator().next() );
-    }
-
     @Value
-    private static class SimpleSVDValueObject {
+    public static class SimpleSVDValueObject {
         /**
          * Order same as the rows of the v matrix.
          */
@@ -1528,6 +1574,14 @@ public class DatasetsWebService {
             this.inferredTerms = inferredTerms.stream()
                     .map( t -> new CharacteristicValueObject( t.getLabel(), t.getUri() ) )
                     .collect( Collectors.toList() );
+        }
+    }
+
+    private List<Long> sliceIds( List<Long> ids, int offset, int limit ) {
+        if ( offset < ids.size() ) {
+            return ids.subList( offset, Math.min( offset + limit, ids.size() ) );
+        } else {
+            return Collections.emptyList();
         }
     }
 }
